qemu: launch spice client when available
'pmbootstrap qemu' will look for spice clients in the user's PATH and
run qemu using the QXL driver if possible.

Currently supported spice clients are 'spicy' and 'remote-viewer' but
adding support for more clients can be easily done.

It is possible to specify the --no-spice argument to force not using QXL.

qemu with qxl support will run on port 8077/tcp, which doesn't belong to
any well-known service and represents 'PM' in decimal.