weston: add a patch to force FBIOPUT on Qualcomm devices
Qualcomm's framebuffer is designed for double buffering, and only updates the screen
after an FBIOPUT ioctl is submitted to flip the buffer. Detect this to force Weston
to send this ioctl at the end of every frame even when single buffering.

This should be equivalent to the
`cat /sys/class/graphics/fb0/modes > /sys/class/graphics/fb0/mode`
hack.