pmbootstrap: support native cross compile for Autoconf packages
This commit adds support for cross-compiling Autotools-based packages
without distcc and qemu. This should be faster and more reliable.

See #659 for more information.

This installs dependencies in the foreign arch chroot, then bind mounts
the foreigh chroot into the native chroot for a cross compile sysroot.
Finally, this sets env variables to trigger a cross compile based on
https://dev.alpinelinux.org/~tteras/bootstrap/abuild-crossbuild-aarch64.conf

Currently chocolate-doom is the only Autotools package whitelisted to build
using this configuration, and the executable it builds on aarch64 seems to
at least start (inside the aarch64 build chroot, not on device)