pmbootstrap: native-cross: install makedepends in both native & target
This allows building Weston in native cross compile mode.

This installs makedepends_host into the target chroot and makedepends_build
into the native chroot. If these vars are not found in the APKBUILD,
it installs makedepends into both.

Note that strict mode isn't supported yet, as Abuild is used to install
dependencies there, and I'm not sure how to allow it to install into the
target.